consolidate email mocks to use one mock from the same mock file
