+++ test/mock/data.ts
@@ -35,7 +35,7 @@ export const MOCK_JWT = 'ey32fakejwt3957t7u2';
 
 export const MOCK_SUBMISSION_ID = '123456789';
 
-export const MOCK_SUBMISSION_EMAIL_ADDRESS = 'someemail@fake.com';
+export const MOCK_EMAIL = 'test@example.com';
 
 export const MOCK_DYNAMODB_RECORD = {
     'id': {
@@ -76,7 +76,7 @@ export const MOCK_DYNAMODB_RECORD = {
                 ]
             }
         },
-        submission_email_address: { S: MOCK_SUBMISSION_EMAIL_ADDRESS }
+        submission_email_address: { S: MOCK_EMAIL }
     }
 };
